+++ extension/nemo_rsync_extension.py
@@ -9,16 +9,42 @@ class NemoRsyncExtension(GObject.GObject, Nemo.MenuProvider):
 
         print("get_file_items() called")
 
-        item = Nemo.MenuItem(
-            name="NemoRsync::Test",
-            label="Rsync Test",
-            tip="Test Nemo Rsync extension"
+        rsync_item = Nemo.MenuItem(
+            name="NemoRsync::Root",
+            label="Rsync",
+            tip="Rsync operations"
         )
 
-        item.connect("activate", self.on_test)
+        submenu = Nemo.Menu()
 
-        return [item]
+        local_item = Nemo.MenuItem(
+            name="NemoRsync::Local",
+            label="Rsync to...",
+            tip="Copy files using rsync"
+        )
 
-    def on_test(self, menu, *args):
+        ssh_item = Nemo.MenuItem(
+            name="NemoRsync::SSH",
+            label="Rsync to SSH...",
+            tip="Copy files over SSH"
+        )
 
-        print("Hello from Nemo Rsync!")
+        local_item.connect("activate", self.on_rsync)
+
+        ssh_item.connect("activate", self.on_rsync_ssh)
+
+        submenu.append_item(local_item)
+        submenu.append_item(ssh_item)
+
+        rsync_item.set_submenu(submenu)
+
+        return [rsync_item]
+
+    def on_rsync(self, menu, *args):
+
+        print("Local rsync selected")
+
+
+    def on_rsync_ssh(self, menu, *args):
+
+        print("SSH rsync selected")